Split C++ provider into interface and implementation.
