+++ BUILD.gn
@@ -1049,6 +1049,10 @@ config("grpc_config") {
         "src/core/lib/json/json_writer.cc",
         "src/core/lib/matchers/matchers.cc",
         "src/core/lib/matchers/matchers.h",
+        "src/core/lib/security/authorization/authorization_engine.h",
+        "src/core/lib/security/authorization/authorization_policy_provider.h",
+        "src/core/lib/security/authorization/authorization_policy_provider_vtable.cc",
+        "src/core/lib/security/authorization/evaluate_args.h",
         "src/core/lib/security/context/security_context.cc",
         "src/core/lib/security/context/security_context.h",
         "src/core/lib/security/credentials/alts/alts_credentials.cc",
@@ -1488,10 +1492,6 @@ config("grpc_config") {
         "include/grpcpp/support/time.h",
         "include/grpcpp/support/validate_service_config.h",
         "include/grpcpp/xds_server_builder.h",
-        "src/core/lib/security/authorization/authorization_engine.h",
-        "src/core/lib/security/authorization/authorization_policy_provider.h",
-        "src/core/lib/security/authorization/authorization_policy_provider_vtable.cc",
-        "src/core/lib/security/authorization/evaluate_args.h",
         "src/cpp/client/channel_cc.cc",
         "src/cpp/client/client_callback.cc",
         "src/cpp/client/client_context.cc",
