+++ grpc.gemspec
@@ -1023,6 +1023,10 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/lib/profiling/basic_timers.cc )
   s.files += %w( src/core/lib/profiling/stap_timers.cc )
   s.files += %w( src/core/lib/profiling/timers.h )
+  s.files += %w( src/core/lib/security/authorization/authorization_engine.h )
+  s.files += %w( src/core/lib/security/authorization/authorization_policy_provider.h )
+  s.files += %w( src/core/lib/security/authorization/authorization_policy_provider_vtable.cc )
+  s.files += %w( src/core/lib/security/authorization/evaluate_args.h )
   s.files += %w( src/core/lib/security/context/security_context.cc )
   s.files += %w( src/core/lib/security/context/security_context.h )
   s.files += %w( src/core/lib/security/credentials/alts/alts_credentials.cc )
