+++ grpc.gyp
@@ -886,6 +886,7 @@
         'src/core/lib/json/json_util.cc',
         'src/core/lib/json/json_writer.cc',
         'src/core/lib/matchers/matchers.cc',
+        'src/core/lib/security/authorization/authorization_policy_provider_vtable.cc',
         'src/core/lib/security/context/security_context.cc',
         'src/core/lib/security/credentials/alts/alts_credentials.cc',
         'src/core/lib/security/credentials/alts/check_gcp_environment.cc',
@@ -1327,6 +1328,7 @@
         'src/core/lib/json/json_reader.cc',
         'src/core/lib/json/json_util.cc',
         'src/core/lib/json/json_writer.cc',
+        'src/core/lib/security/authorization/authorization_policy_provider_vtable.cc',
         'src/core/lib/slice/b64.cc',
         'src/core/lib/slice/percent_encoding.cc',
         'src/core/lib/slice/slice.cc',
@@ -1394,7 +1396,6 @@
         'grpc',
       ],
       'sources': [
-        'src/core/lib/security/authorization/authorization_policy_provider_vtable.cc',
         'src/cpp/client/channel_cc.cc',
         'src/cpp/client/client_callback.cc',
         'src/cpp/client/client_context.cc',
@@ -1528,7 +1529,6 @@
         'grpc_unsecure',
       ],
       'sources': [
-        'src/core/lib/security/authorization/authorization_policy_provider_vtable.cc',
         'src/cpp/client/channel_cc.cc',
         'src/cpp/client/client_callback.cc',
         'src/cpp/client/client_context.cc',
