+++ include/grpcpp/security/authorization_policy_provider.h
@@ -15,9 +15,6 @@
 #ifndef GRPCPP_SECURITY_AUTHORIZATION_POLICY_PROVIDER_H
 #define GRPCPP_SECURITY_AUTHORIZATION_POLICY_PROVIDER_H
 
-#include <grpc/status.h>
-#include <grpcpp/impl/codegen/grpc_library.h>
-
 #include <memory>
 
 // TODO(yihuazhang): remove the forward declaration here and include
@@ -35,30 +32,6 @@ class AuthorizationPolicyProviderInterface {
  public:
   virtual ~AuthorizationPolicyProviderInterface() = default;
   virtual grpc_authorization_policy_provider* c_provider() = 0;
-};
-
-// Implementation obtains authorization policy from static string. This provider
-// will always return the same authorization engines.
-class StaticDataAuthorizationPolicyProvider
-    : public AuthorizationPolicyProviderInterface {
- public:
-  static std::shared_ptr<StaticDataAuthorizationPolicyProvider> Create(
-      const std::string& authz_policy, grpc::Status* status);
-
-  // Use factory method "Create" to create an instance of
-  // StaticDataAuthorizationPolicyProvider.
-  explicit StaticDataAuthorizationPolicyProvider(
-      grpc_authorization_policy_provider* provider)
-      : c_provider_(provider) {}
-
-  ~StaticDataAuthorizationPolicyProvider() override;
-
-  grpc_authorization_policy_provider* c_provider() override {
-    return c_provider_;
-  }
-
- private:
-  grpc_authorization_policy_provider* c_provider_ = nullptr;
 };
 
 }  // namespace experimental
