+++ src/python/grpcio/grpc_core_dependencies.py
@@ -498,6 +498,7 @@ CORE_SOURCE_FILES = [
     'src/core/lib/matchers/matchers.cc',
     'src/core/lib/profiling/basic_timers.cc',
     'src/core/lib/profiling/stap_timers.cc',
+    'src/core/lib/security/authorization/authorization_policy_provider_vtable.cc',
     'src/core/lib/security/context/security_context.cc',
     'src/core/lib/security/credentials/alts/alts_credentials.cc',
     'src/core/lib/security/credentials/alts/check_gcp_environment.cc',
